+++ packages/cedar/src/flatten/flatten.ts
@@ -7,28 +7,31 @@ export interface IFeatureSet {
 }
 
 function flattenFeature(feature: IFeature, categoryField?: string) {
-  const attributes = feature.attributes
+  const attributes = feature.attributes ? feature.attributes : feature
   if (categoryField) {
     // render function expects a property named 'categoryField'
     // TODO: what if attributes already has a 'categoryField' property
     attributes['categoryField'] = attributes[categoryField] || ''
   }
-  return feature.attributes
+  return attributes
 }
 
 // just return an array of flattened features
 export function flattenFeatureSet(featureSet: IFeatureSet, categoryField?) {
+  if (Array.isArray(featureSet)) {
+    return featureSet.map((feature) => (flattenFeature(feature, categoryField)))
+  }
   return featureSet.features.map((feature) => (flattenFeature(feature, categoryField)))
 }
 
-// TODO: don't export this once we're sure we've got the logic right
-export function buildIndex(joinKeys: string[], featureSets: any[]) {
+function buildIndex(joinKeys: string[], featureSets: any[]) {
   const index = {}
   featureSets.forEach((featureSet, i) => {
+    if (featureSet && Array.isArray(featureSet)) { featureSet = { features: featureSet } }
     featureSet.features.forEach((feature) => {
       // the column that this dataset should be joined on
       const joinKey = joinKeys[i]
-      const joinValue = feature.attributes[joinKeys[i]]
+      const joinValue = feature.attributes ? feature.attributes[joinKeys[i]] : feature[joinKeys[i]]
       if (index[joinValue] === undefined) {
         index[joinValue] = []
       }
@@ -60,7 +63,7 @@ export function flattenFeatureSets(featureSets: any[], joinKeys: any[]) {
     const features = index[uniqueValue]
     const row = { categoryField: features[0].attributes[key] || '' }
     features.forEach((feature, i) => {
-      const attributes = feature.attributes
+      const attributes = feature.attributes ? feature.attributes : feature
       Object.keys(attributes).forEach((attrKey) => {
         // append the feature's index to the row's column name
         const col = `${attrKey}_${i}`